requirements.txt
Flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
python-whois==0.8.0
reportlab==4.0.0
beautifulsoup4==4.12.2
flask-login==0.6.3
pytest==7.4.0
celery[redis]==5.3.4
